Applies suggestions from code review
